+++ releasenotes.md
@@ -1,4 +1,4 @@
 Whats New:
 ----------------------
-This is a bug fix release
 Include compositions in Where Used page
+Add solvents and carriers as a separate category